Extract ShareCache into a subpackage
